nslab2/requirements.txt
# pip3 install -r requirements.txt
cryptography
numpy
Pillow
termcolor
